+++ app/routes/marketplace/index.js
@@ -1,5 +1,5 @@
 import createRedirectRoute from "balanced-dashboard/utils/create-redirect-route";
 
-var MarketplaceIndexRoute = createRedirectRoute('marketplace.transactions');
+var MarketplaceIndexRoute = createRedirectRoute('marketplace.orders');
 
 export default MarketplaceIndexRoute;