+++ app/views/detail-views/summary-sections/credit-summary-section.js
@@ -4,8 +4,8 @@ var CreditSummarySectionView = SummarySectionView.extend({
 	statusText: Ember.computed.alias('model.status_description'),
 
 	linkedResources: function() {
-		return this.resourceLinks("model.order", "model.reversals", "model.customer", "model.destination");
-	}.property("model.order", "model.reversals", "model.reversals.length", "model.customer", "model.destination")
+		return this.resourceLinks("model.customer", "model.destination");
+	}.property("model.customer", "model.destination")
 });
 
 export default CreditSummarySectionView;