+++ app/views/marketplace/orders.js
@@ -1,0 +1,6 @@
+var MarketplaceOrdersView = Ember.View.extend({
+	layoutName: "marketplace/payments-layout",
+	templateName: "marketplace/orders",
+});
+
+export default MarketplaceOrdersView;