+++ app/views/results/grouped-transactions-table.js
@@ -1,0 +1,9 @@
+import Ember from "ember";
+
+var GroupedTransactionsTableView = Ember.View.extend({
+	layoutName: 'results/grouped-transactions-table-layout',
+	tagName: 'table',
+	classNames: ["items", "grouped-transactions"],
+});
+
+export default GroupedTransactionsTableView;