+++ app/views/tables/cells/linked-card.js
@@ -1,0 +1,9 @@
+import Ember from "ember";
+
+var LinkedCardView = Ember.View.extend({
+	tagName: "div",
+	templateName: "tables/cells/linked-card",
+	classNameBindings: [":card", ":col-md-3", "item.isOverdue:overdue"],
+});
+
+export default LinkedCardView;